requirements.txt
Flask>=2.2.2
Flask-SQLAlchemy>=2.5.1
flask-bootstrap>=3.3.7
Flask-WTF>=1.0.0
wtforms>=3.0.0
PyMySQL>=1.1.0
gunicorn>=20.1.0 
Werkzeug>=2.2
python-dotenv>=1.0.0